Move non-blocking to ZIOSpec
